Move schemas and entity types
