+++ app/components/Entity/EntityRow/Wrapper.tsx
@@ -4,7 +4,8 @@ import { memo, ReactNode, MouseEvent, useContext } from 'react';
 /* Internal dependencies */
 import { Link } from '~/components/common/Link';
 import { PlayerContext } from '~/context/playerContext/playerContextProvider';
-import { Entity, entityMode, AudioType } from '../types';
+import { Audio } from '~/configs/types';
+import { Entity, entityMode } from '../types';
 import { getID } from '../utils';
 
 interface WrapperProps {
@@ -30,7 +31,7 @@ const Wrapper = ({
     const play = (e: MouseEvent) => {
       e.preventDefault();
       if (setCurrent) {
-        setCurrent(data as AudioType);
+        setCurrent(data as Audio);
       }
     };
