+++ app/components/RegisterSuggestion/index.tsx
@@ -1,0 +1,43 @@
+/* External dependencies */
+import React from 'react';
+import { useNavigate } from 'react-router-dom';
+
+/* Internal dependencies */
+import { PrimaryButton } from '~/components/common/Button';
+
+const RegisterSuggestion = () => {
+  const navigate = useNavigate();
+
+  const goToRegister = () => {
+    navigate('/register');
+  };
+
+  return (
+    <>
+      <h3>Not a member yet?</h3>
+      <ul>
+        <li>
+          <span>⚡️</span>
+          <span>You get 2 month free trial.</span>
+        </li>
+        <li>
+          <span>😇</span>
+          <span>You support music since Muzikie pays a whopping 80% share to artists.</span>
+        </li>
+        <li>
+          <span>💰</span>
+          <span>You receive 10% share of ad-supported profit.</span>
+        </li>
+      </ul>
+      <PrimaryButton
+        className='registerButton'
+        theme="white"
+        onClick={goToRegister}
+      >
+        Register
+      </PrimaryButton>
+    </>
+  );
+};
+
+export default RegisterSuggestion;
