+++ app/components/UserDiscography/types.ts
@@ -1,6 +1,6 @@
-import { CollectionType, AudioType } from '../Entity/types';
+import { Collection, Audio } from '~/configs/types';
 
 export interface UserDiscographyProps {
-  audios: AudioType[];
-  collections: CollectionType[];
+  audios: Audio[];
+  collections: Collection[];
 }
