+++ app/configs/schemas.ts
@@ -1,28 +1,3 @@
-export const MODULES = {
-  TOKEN: 'token',
-  AUTH: 'auth',
-  COLLECTION: 'collection',
-  AUDIO: 'audio',
-  SUBSCRIPTION: 'subscription',
-};
-
-export const COMMANDS = {
-  STREAM: 'stream',
-  CREATE: 'create',
-  PURCHASE: 'purchase',
-  TRANSFORM: 'transform',
-  DESTROY: 'destroy',
-  SET_ATTRIBUTES: 'setAttributes',
-  UPDATE_MEMBERS: 'updateMembers',
-};
-
-export const FEEDBACK_MESSAGES = {
-  INVALID_PARAMS: 'Audio parameters were invalid. Please review the data.',
-  BROADCAST_ERROR: 'Error creating audio. Try again.',
-  SUCCESS: 'Audio created successfully',
-  PENDING: 'Wait while audio is being create. (~10 seconds)',
-};
-
 export const AUDIO_CREATE_SCHEMA = {
   $id: 'audio/create',
   title: 'CreateAsset transaction asset for audio module',
@@ -171,4 +146,3 @@ export const SUBSCRIPTION_PURCHASE_SCHEMA = {
     },
   },
 };
-
