+++ app/entry.server.tsx
@@ -4,7 +4,7 @@ import type { EntryContext } from '@remix-run/node';
 import { Response } from '@remix-run/node';
 import { RemixServer } from '@remix-run/react';
 import { renderToPipeableStream } from 'react-dom/server';
-import { ABORT_DELAY } from '~/constants/app';
+import { ABORT_DELAY } from '~/configs/app';
 
 export default function handleRequest(
   request: Request,
