+++ app/models/entity.client.ts
@@ -1,9 +1,6 @@
-import type {
-  CollectionType,
-  AudioType,
-  Entity,
-} from '~/components/Entity/types';
-import { API_URLS } from '~/constants/api';
+import type { Entity } from '~/components/Entity/types';
+import type { Collection, Audio } from '~/configs/types';
+import { API_URLS } from '~/configs/api';
 
 export type SearchResultType = {
   [key: string]: Entity[];
@@ -17,14 +14,14 @@ const post = (url: string, body: any) => fetch(
   }
 ).then((res) => res.json()).then(res => res.data).catch(console.log);
 
-export async function postCollection(json: Entity, file: File): Promise<CollectionType> {
+export async function postCollection(json: Entity, file: File): Promise<Collection> {
   const data = new FormData();
   data.append('file', file);
   data.append('data', JSON.stringify(json));
   return post(`${API_URLS.STREAMER}/api/v1/collections`, data);
 }
 
-export async function postAudio(json: Entity, file: File): Promise<AudioType> {
+export async function postAudio(json: Entity, file: File): Promise<Audio> {
   const data = new FormData();
   data.append('file', file);
   data.append('data', JSON.stringify(json));
